bug fixed

+++ src/main/java/dao/ArticleDAO.java
@@ -2,5 +2,5 @@ package dao;
 
 import model.Article;
 
-public interface ArticleDAO extends GenericDAO<Article, Long> {
+public interface ArticleDAO extends GenericDAO<Article, Integer> {
 }
